ext/imap/tests/*mutf7*.phpt: update for missing utf8_to_mutf7() (#11654)
Our ./configure script (ext/imap/config.m4) checks for the
utf8_to_mutf7() function in the c-client library, and defines
HAVE_IMAP_MUTF7 only if it exists. The two corresponding PHP
functions are disabled when it does not, but their tests do
not account for that. Here we add two SKIPIFs to hand that
scenario.